Lab4/steamshp_project/api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchvision==0.22.1+cpu
torchaudio==2.7.1+cpu
fastapi==0.115.13
uvicorn==0.34.3
transformers==4.52.4
requests==2.32.3
psycopg2-binary==2.9.10
sentencepiece==0.2.0
